refactor:password and profile actions separated into individual controllers
